refactor: extract RowEditingCoordinator from MainContentCoordinator
