+++ TablePro/Views/Main/MainContentCoordinator.swift
@@ -102,6 +102,7 @@ final class MainContentCoordinator {
     @ObservationIgnored private(set) var filterCoordinator: FilterCoordinator!
     @ObservationIgnored private(set) var queryExecutionCoordinator: QueryExecutionCoordinator!
     @ObservationIgnored private(set) var paginationCoordinator: PaginationCoordinator!
+    @ObservationIgnored private(set) var rowEditingCoordinator: RowEditingCoordinator!
 
     /// Stable identifier for this coordinator's window (set by MainContentView on appear)
     var windowId: UUID?
@@ -399,6 +400,7 @@ final class MainContentCoordinator {
         self.filterCoordinator = FilterCoordinator(parent: self)
         self.queryExecutionCoordinator = QueryExecutionCoordinator(parent: self)
         self.paginationCoordinator = PaginationCoordinator(parent: self)
+        self.rowEditingCoordinator = RowEditingCoordinator(parent: self)
 
         Self.lifecycleLogger.info(
             "[open] MainContentCoordinator.init done connId=\(connection.id, privacy: .public) elapsedMs=\(Int(Date().timeIntervalSince(initStart) * 1_000))"
